separate window class from parallel renderrer
